Separate module for modular solver DSL
Addresses https://github.com/haskell/cabal/pull/2500#commitcomment-10817840.
